Refactoring, added numpy binary format
