+++ src/pymodaq/daq_utils/h5node_exporters.py
@@ -1,12 +1,76 @@
 # Standard imports
 from abc import ABCMeta, abstractmethod
+from typing import Callable
 
 # 3rd party imports
 import numpy as np
 
 # project imports
-from pymodaq.daq_utils.h5modules import H5BrowserUtil,H5Backend
+from pymodaq.daq_utils.h5backend import H5Backend
+from pymodaq.daq_utils.daq_utils import set_logger, get_module_name
 
+logger = set_logger(get_module_name(__file__))
+
+class ExporterFactory:
+    """The factory class for creating executors"""
+
+    exporters_registry = {}
+    file_filters = {}
+
+    @classmethod
+    def register_exporter(cls) -> Callable:
+        """Class decorator method to register exporter class to the internal registry. Must be used as
+        decorator above the definition of an H5Exporter class. H5Exporter must implement specific class
+        attributes and methods, see definition: h5node_exporter.H5Exporter
+
+        See h5node_exporter.H5txtExporter and h5node_exporter.H5txtExporter for usage examples.
+
+        returns:
+            the exporter class
+        """
+
+        def inner_wrapper(wrapped_class) -> Callable:
+            extension = wrapped_class.FORMAT_EXTENSION
+            #Warn if overriding existing exporter
+            if extension in cls.exporters_registry:
+                logger.warning(f"Exporter for the .{extension} format already exists and will be replaced")
+
+            #Register extension
+            cls.exporters_registry[extension] = wrapped_class
+            cls.file_filters[extension] = wrapped_class.FORMAT_DESCRIPTION
+            #Return wrapped_class
+            return wrapped_class
+
+        #Return decorated function
+        return inner_wrapper
+
+    @classmethod
+    def create_exporter(cls, extension: str):
+        """Factory command to create the exporter object.
+
+            This method gets the appropriate executor class from the registry
+            and instantiates it.
+
+            Args:
+                extension (str): the extension of the file that will be exported
+
+            returns:
+                an instance of the executor created
+        """
+        if extension not in cls.exporters_registry:
+            raise ValueError(f".{extension} is not a supported file format.")
+
+        exporter_class = cls.exporters_registry[extension]
+
+        exporter = exporter_class()
+
+        return exporter
+
+    @classmethod
+    def get_file_filters(cls):
+        """Create the file filters string"""
+        tmplist = [f"{v} (*.{k})" for k,v in cls.file_filters.items()]
+        return ";;".join(tmplist)
 
 class H5Exporter(metaclass=ABCMeta):
     """Base class for an exporter. """
@@ -35,7 +99,7 @@ class H5Exporter(metaclass=ABCMeta):
         """Abstract method to save a .h5 node to a file"""
         pass
 
-@H5BrowserUtil.register_exporter()
+@ExporterFactory.register_exporter()
 class H5h5Exporter(H5Exporter):
     """ Exporter object for saving nodes as single h5 files"""
 
@@ -55,7 +119,7 @@ class H5h5Exporter(H5Exporter):
         new_file.h5file.remove_node('/Raw_datas', recursive=True)
         new_file.close_file()
 
-@H5BrowserUtil.register_exporter()
+@ExporterFactory.register_exporter()
 class H5txtExporter(H5Exporter):
     """ Exporter object for saving nodes as txt files"""
 
@@ -72,7 +136,6 @@ class H5txtExporter(H5Exporter):
                 np.savetxt(filename, data, '%s', '\t')
             else:
                 np.savetxt(filename, data, '%.6e', '\t')
-
         elif 'GROUP' in node.attrs['CLASS']:
             data_tot = []
             header = []
@@ -99,7 +162,7 @@ class H5txtExporter(H5Exporter):
             data_trans = np.array(list(zip(*data_tot)), dtype=dtypes)
             np.savetxt(filename, data_trans, fmts, '\t', header='#' + '\t'.join(header))
 
-@H5BrowserUtil.register_exporter()
+@ExporterFactory.register_exporter()
 class H5asciiExporter(H5Exporter):
     """ Exporter object for saving nodes as txt files"""
 
@@ -146,3 +209,19 @@ class H5asciiExporter(H5Exporter):
             data_trans = np.array(list(zip(*data_tot)), dtype=dtypes)
 
             np.savetxt(filename, data_trans, fmts, '\t', header='#' + '\t'.join(header))
+
+@ExporterFactory.register_exporter()
+class H5npyExporter(H5Exporter):
+    """ Exporter object for saving nodes as npy files"""
+
+    FORMAT_DESCRIPTION = "Binary NumPy format"
+    FORMAT_EXTENSION = "npy"
+
+    def _export_data(self, node, filename) -> None:
+        """Export the node as a numpy binary file format"""
+        if 'ARRAY' in node.attrs['CLASS']:
+            data = node.read()
+            if not isinstance(data, np.ndarray):
+                data = np.array(data)
+
+            np.save(filename, data)
